Expressltexx: Add custom RIL class
Apparently we have to use and maintain our own RIL class now
for every device that will be using KitKat RIL blobs.

Completely unnecessary and bloated imo, but hey - if that's how
the flow goes these days, then by all means. It's not like we
have too little of it already. More is less.

Change-Id: Idb9a518e741fcc0c85dac65daed89c02dc71eb76